Revert "Add support for Redux"
