--- client/app/actions/token.js
@@ -1,9 +1,0 @@
-import { TOKEN } from '../constants/token'
-function addToken (token) {
-  return {
-    type: TOKEN.ADD_TOKEN,
-    payload: token
-  }
-}
-
-export default { addToken }
