--- client/app/constants/token.js
@@ -1,3 +1,0 @@
-export const TOKEN = {
-  ADD_TOKEN: 'ADD_TOKEN'
-}
