+++ client/app/core/config.js
@@ -1,6 +1,4 @@
 /* eslint angular/window-service: "off" */
-import { RootReducer } from '../reducers'
-import { autoRehydrate } from 'redux-persist'
 
 var DEVEL_DOMAINS = [
   'localhost',
@@ -9,10 +7,9 @@ var DEVEL_DOMAINS = [
 ]
 
 var isDevel = window._.includes(DEVEL_DOMAINS, window.location.hostname)
-const hasDevTools = angular.isDefined(window.__REDUX_DEVTOOLS_EXTENSION__)
 
 /** @ngInject */
-export function configure ($logProvider, $compileProvider, $qProvider, $ngReduxProvider) {
+export function configure ($logProvider, $compileProvider, $qProvider) {
   $logProvider.debugEnabled(isDevel)
   $compileProvider.debugInfoEnabled(isDevel)
 
@@ -20,12 +17,4 @@ export function configure ($logProvider, $compileProvider, $qProvider, $ngReduxP
   $compileProvider.preAssignBindingsEnabled(true)
 
   $qProvider.errorOnUnhandledRejections(false)
-
-  const storeEnhancers = [
-    autoRehydrate()
-  ]
-  if (hasDevTools) {
-    storeEnhancers.push(window.__REDUX_DEVTOOLS_EXTENSION__())
-  }
-  $ngReduxProvider.createStoreWith(RootReducer, [], storeEnhancers)
 }
