--- client/app/reducers/index.js
@@ -1,6 +1,0 @@
-import { combineReducers } from 'redux'
-import { TokenReducer } from './token'
-
-export const RootReducer = combineReducers({
-  token: TokenReducer
-})
