+++ package.json
@@ -93,7 +93,6 @@
     "merge": "1.2.0",
     "moment-timezone": "0.5.13",
     "ng-annotate-loader": "0.6.1",
-    "ng-redux": "3.5.2",
     "ngprogress": "VictorBjelkholm/ngProgress#v1.1.3",
     "ngstorage": "0.3.11",
     "ngtemplate-loader": "2.0.1",
@@ -104,8 +103,6 @@
     "patternfly-timeline": "patternfly/patternfly-timeline#master-dist",
     "postcss-loader": "2.0.8",
     "q": "1.5.1",
-    "redux": "3.7.2",
-    "redux-persist": "4.10.1",
     "request": "2.83.0",
     "rxjs": "5.5.0",
     "sass-loader": "6.0.6",
